removed hard coded password
git-svn-id: file:///home/svn/shoola-omero/trunk@3773 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/env/data/OmeroServiceImpl.java
@@ -354,10 +354,15 @@ class OmeroServiceImpl
         if (child == null) 
             throw new IllegalArgumentException("The child cannot be null.");
         if (!(child instanceof ImageData)) {
-            IObject object = ModelMapper.removeIObject(child.asIObject(),
-                    parent.asIObject());
-            PojoMapper.asDataObject(gateway.updateObject(object,
-                    (new PojoOptions()).map()));
+            if (parent == null) { //top container
+                PojoMapper.asDataObject(gateway.updateObject(child.asIObject(),
+                        (new PojoOptions()).map()));
+            } else {
+                IObject object = ModelMapper.removeIObject(child.asIObject(),
+                        parent.asIObject());
+                PojoMapper.asDataObject(gateway.updateObject(object,
+                        (new PojoOptions()).map()));
+            }
         } else {
             IObject p = parent.asIObject();
             Image img = child.asImage();
